# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediAsk API",
    "version": "1.0.0",
    "description": "智能医疗辅助问诊系统 - 按模块组织"
  },
  "servers": [
    {
      "url": "http://localhost:8989",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/schedule-templates": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "创建排班模板",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1ScheduleTemplates"
      }
    },
    "/api/v1/schedule-templates/{templateId}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "查询排班模板详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1ScheduleTemplatesByTemplateId"
      },
      "put": {
        "tags": [
          "schedule"
        ],
        "summary": "更新排班模板",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "putApiV1ScheduleTemplatesByTemplateId"
      }
    },
    "/api/v1/schedule-templates/{templateId}/publish": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "发布排班模板",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1ScheduleTemplatesByTemplateIdPublish"
      }
    },
...
    "/api/v1/schedule-rule-profiles": {
      "post": {
        "tags": [
          "schedule-rule-profile"
        ],
        "summary": "创建规则配置草稿版本",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRuleProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRuleProfileResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiV1ScheduleRuleProfiles"
      }
    },
    "/api/v1/schedule-rule-profiles/{profileId}": {
      "get": {
        "tags": [
          "schedule-rule-profile"
        ],
        "summary": "查询规则配置详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "规则配置ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRuleProfileResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1ScheduleRuleProfilesByProfileId"
      }
    },
    "/api/v1/schedule-rule-profiles/versions": {
      "get": {
        "tags": [
          "schedule-rule-profile"
        ],
        "summary": "查询规则配置版本列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "departmentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "profileCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRuleProfileVersionListResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1ScheduleRuleProfilesVersions"
      }
    },
    "/api/v1/schedule-rule-profiles/{profileId}/publish": {
      "post": {
        "tags": [
          "schedule-rule-profile"
        ],
        "summary": "发布规则配置",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "规则配置ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRuleProfilePublishResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiV1ScheduleRuleProfilesByProfileIdPublish"
      }
    },
    "/api/v1/schedule-rule-profiles/{profileId}/rollback": {
      "post": {
        "tags": [
          "schedule-rule-profile"
        ],
        "summary": "回滚发布到指定规则版本",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "规则配置ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRuleProfilePublishResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiV1ScheduleRuleProfilesByProfileIdRollback"
      }
    },
...
    "/api/v1/schedules/{scheduleId}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "查询排班详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1SchedulesByScheduleId"
      },
      "delete": {
        "tags": [
          "schedule"
        ],
        "summary": "删除排班",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "是否强制删除（取消关联预约）",
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "deleteApiV1SchedulesByScheduleId"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateScheduleRuleProfileRequest": {
        "type": "object",
        "required": [
          "departmentId",
          "profileCode",
          "profileName",
          "constraintDslJson"
        ],
        "properties": {
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "profileCode": {
            "type": "string"
          },
          "profileName": {
            "type": "string"
          },
          "constraintDslJson": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ScheduleRuleProfileResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleRuleProfileDTO"
          }
        }
      },
      "ScheduleRuleProfileDTO": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "profileCode": {
            "type": "string"
          },
          "profileName": {
            "type": "string"
          },
          "versionNo": {
            "type": "integer",
            "format": "int32"
          },
          "profileStatus": {
            "type": "string"
          },
          "constraintDslJson": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ScheduleRuleProfileVersionListResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleRuleProfileVersionDTO"
            }
          }
        }
      },
      "ScheduleRuleProfileVersionDTO": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "profileCode": {
            "type": "string"
          },
          "profileName": {
            "type": "string"
          },
          "versionNo": {
            "type": "integer",
            "format": "int32"
          },
          "profileStatus": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ScheduleRuleProfilePublishResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleRuleProfilePublishResultDTO"
          }
        }
      },
      "ScheduleRuleProfilePublishResultDTO": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "integer",
            "format": "int64"
          },
          "profileCode": {
            "type": "string"
          },
          "versionNo": {
            "type": "integer",
            "format": "int32"
          },
          "profileStatus": {
            "type": "string"
          },
          "dslVersion": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "CreateScheduleTemplateRequest": {
        "type": "object",
        "required": [
          "doctorId",
          "templateName",
          "effectiveStartDate",
          "effectiveEndDate",
          "cancelDeadlineMinutes",
          "defaultFee",
          "rules"
        ],
        "properties": {
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "templateName": {
            "type": "string"
          },
          "effectiveStartDate": {
            "type": "string",
            "format": "date"
          },
          "effectiveEndDate": {
            "type": "string",
            "format": "date"
          },
          "cancelDeadlineMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "defaultFee": {
            "type": "number",
            "format": "double"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleTemplateRuleRequest"
            }
          }
        }
      },
...
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "traceId": {
            "type": "string",
            "description": "链路追踪ID"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "响应时间戳（毫秒）"
          }
        },
        "example": {
          "code": 0,
          "msg": "success",
          "data": {}
        }
      },
...
      "ScheduleTemplateRuleRequest": {
        "type": "object",
        "required": [
          "weekday",
          "timePeriodCode",
          "periodStartTime",
          "periodEndTime",
          "slotDurationMinutes",
          "slotCapacity"
        ],
        "properties": {
          "weekday": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 7
          },
          "timePeriodCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 3
          },
          "periodStartTime": {
            "type": "string",
            "format": "time"
          },
          "periodEndTime": {
            "type": "string",
            "format": "time"
          },
          "slotDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "minimum": 5
          },
          "slotCapacity": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "fee": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "UpdateScheduleTemplateRequest": {
        "type": "object",
        "required": [
          "templateName",
          "effectiveStartDate",
          "effectiveEndDate",
          "cancelDeadlineMinutes",
          "defaultFee",
          "rules"
        ],
        "properties": {
          "templateName": {
            "type": "string"
          },
          "effectiveStartDate": {
            "type": "string",
            "format": "date"
          },
          "effectiveEndDate": {
            "type": "string",
            "format": "date"
          },
          "cancelDeadlineMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "defaultFee": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleTemplateRuleRequest"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}